requirements.txt
tensorboard>=2.0.0
torch>=1.3.1,<1.6.0
torchvision>=0.4.1,<0.7.0
opencv-python
pandas
Pillow>=6.2.1
scikit-learn>=0.21.3
tqdm>=4.36.1
scikit-image>=0.16.2
matplotlib==3.1.0